--- a/implementations/rust/ockam/ockam/src/workers/echoer.rs
+++ b/implementations/rust/ockam/ockam/src/workers/echoer.rs
@@ -14,9 +14,9 @@ impl Worker for Echoer {
     type Message = String;
 
     async fn handle_message(&mut self, ctx: &mut Context, msg: Routed<String>) -> Result<()> {
-        debug!("Address: {}, Received: {}", ctx.address(), msg);
+        debug!("Address: {}, Received: {:?}", ctx.address(), msg);
 
         // Echo the message body back on its return_route.
-        ctx.send(msg.return_route(), msg.body()).await
+        ctx.send(msg.return_route(), msg.into_body()?).await
     }
 }