--- a/implementations/rust/ockam/ockam_api/src/authenticator/credential_issuer/credential_issuer_worker.rs
+++ b/implementations/rust/ockam/ockam_api/src/authenticator/credential_issuer/credential_issuer_worker.rs
@@ -55,7 +55,9 @@ impl Worker for CredentialIssuerWorker {
         };
 
         let from = secure_channel_info.their_identity_id();
-        let mut dec = Decoder::new(m.as_body());
+        let return_route = m.return_route();
+        let body = m.into_body()?;
+        let mut dec = Decoder::new(&body);
         let req: RequestHeader = dec.decode()?;
         trace! {
             target: "credential_issuer",
@@ -77,6 +79,6 @@ impl Worker for CredentialIssuerWorker {
             _ => Response::unknown_path(&req).to_vec()?,
         };
 
-        c.send(m.return_route(), res).await
+        c.send(return_route, res).await
     }
 }