--- a/implementations/rust/ockam/ockam_api/src/authenticator/direct/direct_authenticator_worker.rs
+++ b/implementations/rust/ockam/ockam_api/src/authenticator/direct/direct_authenticator_worker.rs
@@ -41,7 +41,9 @@ impl Worker for DirectAuthenticatorWorker {
         };
 
         let from = secure_channel_info.their_identity_id();
-        let mut dec = Decoder::new(m.as_body());
+        let return_route = m.return_route();
+        let body = m.into_body()?;
+        let mut dec = Decoder::new(&body);
         let req: RequestHeader = dec.decode()?;
         trace! {
             target: "direct_authenticator",
@@ -98,7 +100,7 @@ impl Worker for DirectAuthenticatorWorker {
             _ => Response::unknown_path(&req).to_vec()?,
         };
 
-        c.send(m.return_route(), res).await?;
+        c.send(return_route, res).await?;
 
         Ok(())
     }