--- a/implementations/rust/ockam/ockam_api/src/authenticator/enrollment_tokens/acceptor_worker.rs
+++ b/implementations/rust/ockam/ockam_api/src/authenticator/enrollment_tokens/acceptor_worker.rs
@@ -43,7 +43,9 @@ impl Worker for EnrollmentTokenAcceptorWorker {
         };
 
         let from = secure_channel_info.their_identity_id();
-        let mut dec = Decoder::new(m.as_body());
+        let return_route = m.return_route();
+        let body = m.into_body()?;
+        let mut dec = Decoder::new(&body);
         let req: RequestHeader = dec.decode()?;
         trace! {
             target: "enrollment_token_acceptor",
@@ -65,6 +67,6 @@ impl Worker for EnrollmentTokenAcceptorWorker {
             }
             _ => Response::unknown_path(&req).to_vec()?,
         };
-        c.send(m.return_route(), res).await
+        c.send(return_route, res).await
     }
 }