--- a/implementations/rust/ockam/ockam_api/src/authenticator/enrollment_tokens/issuer_worker.rs
+++ b/implementations/rust/ockam/ockam_api/src/authenticator/enrollment_tokens/issuer_worker.rs
@@ -45,7 +45,9 @@ impl Worker for EnrollmentTokenIssuerWorker {
         };
 
         let from = secure_channel_info.their_identity_id();
-        let mut dec = Decoder::new(m.as_body());
+        let return_route = m.return_route();
+        let body = m.into_body()?;
+        let mut dec = Decoder::new(&body);
         let req: RequestHeader = dec.decode()?;
         trace! {
             target: "enrollment_token_issuer",
@@ -74,6 +76,6 @@ impl Worker for EnrollmentTokenIssuerWorker {
             }
             _ => Response::unknown_path(&req).to_vec()?,
         };
-        c.send(m.return_route(), res).await
+        c.send(return_route, res).await
     }
 }