--- a/implementations/rust/ockam/ockam_api/src/echoer.rs
+++ b/implementations/rust/ockam/ockam_api/src/echoer.rs
@@ -12,7 +12,7 @@ impl Worker for Echoer {
     #[instrument(skip_all, name = "Echoer::handle_message")]
     async fn handle_message(&mut self, ctx: &mut Context, msg: Routed<Any>) -> Result<()> {
         log::debug!(src = %msg.src_addr(), from = %msg.sender()?, to = %msg.return_route().step()?, "echoing back");
-        ctx.send(msg.return_route(), NeutralMessage::from(msg.take_payload()))
+        ctx.send(msg.return_route(), NeutralMessage::from(msg.into_payload()))
             .await
     }
 }