--- a/implementations/rust/ockam/ockam_api/src/uppercase.rs
+++ b/implementations/rust/ockam/ockam_api/src/uppercase.rs
@@ -9,7 +9,7 @@ impl Worker for Uppercase {
 
     #[instrument(skip_all, name = "Uppercase::handle_message")]
     async fn handle_message(&mut self, ctx: &mut Context, msg: Routed<String>) -> Result<()> {
-        ctx.send(msg.return_route(), msg.body().to_uppercase())
+        ctx.send(msg.return_route(), msg.into_body()?.to_uppercase())
             .await
     }
 }