In this PR, I address the slowest pieces in ockam portals, extra memory copies and slow serialization/deserialization with serde bare.
The Routed no longer contains the decoded body, but gets decoded only when `Routed::into_body()` is called. This avoid an extra copy of the same message both in form of deserialized and serialized form. (this also reduces the peak memory usage).

Every BARE message involved with TCP portals is now manually serialized/deserilized, in specific:
- TCP portal processor/worker 
- Ockam TCP messaging protocol (`TransportMessage`)

TCP Portal message and Secure Channel messages, where possible/beneficial, are  now borrow buffer rather than owning the `Vec<u8>` to avoid extra copies (~zero copy deserialization).

The encoding/decoding primitives for serde were implemented with `no_std` within the `ockam_core` crate.

The secure channel encryptor/decryptor were already using CBOR instead (which is much faster than `serde_cbor`,  manually serializing/deserializing doesn't make much sense), but there is an extra serde encoding of a `Vec<u8>`. To avoid the serde-encoding pass and an extra buffer copy, a hack was implemented. Since the serialization of a `Vec<u8>` in bare is just the length followed the raw buffer, by calculating beforehand the length of the resulting encrypted data we can directly encrypt inside the same buffer. We have to calculate the encrypted length beforehand since the length of the array in bare is variable depending on the value. @SanjoDeundiak I'm NOT sure if we can actually assume the encrypted size is always payload + overhead, and that the overhead is a static number.

```
amd64 laptop - 11th Gen Intel(R) Core(TM) i7-1165G7 @ 2.80GHz

before optimizations:
one node portal: 7.16 GBytes  1.02 Gbits/sec
two nodes portal: 5.73 GBytes   819 Mbits/sec
message latency: 101.066µs

after optimizations:
one node portal: 42.2 GBytes  6.04 Gbits/sec - delta: ~5.9x
two nodes portal: 40.8 GBytes  5.83 Gbits/sec - delta: ~7.1x
message latency: 86.445µs - delta: ~0.855x
```

```
arm64 laptop - Apple M2 Pro 12 Cores

before optimizations:
one node portal: 4.72 GBytes   675 Mbits/sec
two nodes portal: 4.84 GBytes   692 Mbits/sec
message latency: 91.512µs

after optimizations:
one node portal: 7.10 GBytes  1.02 Gbits/sec - delta: ~1.7x
two nodes portal: 7.01 GBytes  1.00 Gbits/sec - delta: ~1.44x
message latency: 82.323µs - delta: ~0.90x
```